Add attribute to force exec transition
Previously command would always use the target configuration, which
might not be ideal if you're building a tool alongside another target,
and therefore get the dbg compilation mode. This allows you to set a new
attribute to always use the exec configuration instead. This generally
assumes that your target and exec configurations for commands will be
the same.