Add transition to force opt for command tools
These tools are using the target configuration because theoretically in
the case you had remote execution on Linux with macOS hosts, you need
the target to match your machine for when its run. But in this specific
case these commands should likely always use the opt configuration, but
you may not want to force that on the command line to avoid analysis
cache trashing. This adds a custom transition to the targets in order to
force opt.