Test rlocation after change of directory
Bazel's runfiles lookup mechanism can handle directory changes, to that
end it looks up the runfiles tree or manifest relative to argv0 or a
preset environment variable.

A common use-case that requires this are executables that are meant to
be run in the workspace root, e.g. BUILD file generation scripts.