Add boolean to run from workspace root (#4)
This is common for tools like linters. Doing the cd themselves is fine, but for some other tools you consume, this allows you to avoid a shim just to do this. For example if you run `flake8` from source, you either have to write a shim or do something like this.