fix: Handle RunEnvironmentInfo (#68)
Executable rules can include a
[RunEnvironmentInfo](https://bazel.build/rules/lib/providers/RunEnvironmentInfo)
provider to set envs. These envs are detected and included with
explicitly defined envs on the target when defining the multirun
commands.

Because bazel allows for this provider, it is nice for rules_multirun to
handle it. I ran into a situation where a 3rd party ruleset uses this as
part of a run action and there is no way to provide the correct value by
wrapping in a `command` as the env is internal to the ruleset.