Fix deadlock with large output (#71)
Specifically in the docs of `process.wait()` there is:

```
Note This will deadlock when using stdout=PIPE or stderr=PIPE and the
child process generates enough output to a pipe such that it blocks
waiting for the OS pipe buffer to accept more data. Use
Popen.communicate() when using pipes to avoid that.
```

Now we read stdout as it comes in and wait for the process afterwards.